use task model

+++ app/Http/Controllers/TasksController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
-use App\Tasklist;
+use App\Task;
 
 class TasksController extends Controller
 {
@@ -15,11 +15,21 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks  = Tasklist::all();
-        
-        return view('tasks.index',[
-            'tasks' => $tasks,
-        ]);
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
     }
 
     /**
@@ -29,7 +39,7 @@ class TasksController extends Controller
      */
     public function create()
     {
-        $task = new Tasklist;
+        $task = new Task;
         
         return view('tasks.create',[
             'task' => $task,
@@ -44,14 +54,19 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
+        // バリデーション
         $request->validate([
-            'status' => 'required|max:10', 
+            'status' => 'required|max:10',
             'content' => 'required|max:255',
         ]);
-        $task = new Tasklist;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
+
+        // 前のURLへリダイレクトさせる
         return redirect('/');
     }
 
@@ -63,10 +78,14 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Tasklist::findOrFail($id);
-        return view('tasks.show',[
-            'task' => $task,    
-        ]);
+        $task = Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show',[
+                'task' => $task,    
+            ]);
+        }else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -77,10 +96,14 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        $task = Tasklist::findOrFail($id);
-        return view(tasks.update,[
-            'task' => $task,
+        $task = Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id) {
+        return view('tasks.edit',[
+            'task' => $task,    
         ]);
+        }else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -93,14 +116,13 @@ class TasksController extends Controller
     public function update(Request $request, $id)
     {
         $request->validate([
-            'status' => 'required|max:10', 
+            'status' => 'required|max:10',
             'content' => 'required|max:255',
         ]);
-        $task = Tasklist::findOrFail($id);
+        $task = Task::findOrFail($id);
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
-        
         return redirect('/');
     }
 
@@ -112,9 +134,15 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Tasklist::findOrFail($id);
-        $task->delete();
-        
+        // idの値で投稿を検索して取得
+        $task = Task::findOrFail($id);
+
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        // 前のURLへリダイレクトさせる
         return redirect('/');
     }
 }
